Added a simple navigation
#6ggwu5[InReview]
