+++ package.json
@@ -8,14 +8,21 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "@react-native-community/masked-view": "0.1.6",
+    "@react-navigation/native": "^5.1.7",
     "expo": "~37.0.3",
     "i": "^0.3.6",
     "react": "~16.9.0",
     "react-dom": "~16.9.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-37.0.1.tar.gz",
-    "react-native-screens": "~2.2.0",
+    "react-native-gesture-handler": "^1.6.1",
+    "react-native-reanimated": "^1.8.0",
+    "react-native-safe-area-context": "^0.7.3",
+    "react-native-screens": "^2.7.0",
     "react-native-vector-icons": "^6.6.0",
     "react-native-web": "~0.11.7",
+    "react-navigation": "^4.3.8",
+    "react-navigation-stack": "^2.3.13",
     "uuidv4": "^6.0.7"
   },
   "devDependencies": {
